fix(tests): move all mocks to folder mocks
all mocked data is now in one folder and the imports are reorganized accordingly
