+++ test/viewmodels/home_view_model_unit_test.dart
@@ -10,11 +10,11 @@ import "package:proxima/services/database/user_repository_service.dart";
 import "package:proxima/services/geolocation_service.dart";
 import "package:proxima/viewmodels/home_view_model.dart";
 
-import "../models/database/post/mock_post_data.dart";
-import "../models/database/user/mock_user_data.dart";
-import "../services/database/mock_post_repository_service.dart";
-import "../services/database/mock_user_repository_service.dart";
-import "../services/mock_geo_location_service.dart";
+import "../mocks/mock_post_data.dart";
+import "../mocks/mock_user_data.dart";
+import "../mocks/mock_post_repository_service.dart";
+import "../mocks/mock_user_repository_service.dart";
+import "../mocks/mock_geo_location_service.dart";
 
 void main() {
   group("Post Overview Provider unit testing", () {
